% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropDB.R
\name{File2DB}
\alias{dropDB}
\title{dropDB function}
\usage{
dropDB(db1, db2, db3, db4, db5)
}
\arguments{
\item{db1}{db2 db3 db4 db5
return delete db table from sql server}
}
\description{
dropDB function
DB 서버상에서 특정 테이블을 삭제합니다다
}
