% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagetypedb.R
\name{pagetypedb}
\alias{pagetypedb}
\title{pagetypedb function}
\usage{
dropDB(db1, db2, db3, db4, db5)
}
\arguments{
\item{db1}{db2 db3 db4 db5
return delete db table from sql server}
}
\description{
dropDB function
}
